refactor: Rename the main gtest testing helper class to DetoursTest
