+++ msvc/corehook-test/DetoursTest.h
